requirements.txt
streamlit>=1.0
pandas>=1.3
python-docx>=0.8
reportlab>=3.6
python-dateutil>=2.8
